+++ src/main/java/ru/vsu/noidle_server/model/domain/RequirementEntity.java
@@ -7,7 +7,6 @@ import ru.vsu.noidle_server.model.converter.SubTypeConverter;
 import ru.vsu.noidle_server.model.converter.TypeConverter;
 
 import javax.persistence.*;
-import java.math.BigInteger;
 import java.util.Collection;
 import java.util.Objects;
 
@@ -42,7 +41,7 @@ public class RequirementEntity {
     private String name;
 
     @Column(name = "value", nullable = false)
-    private BigInteger value;
+    private Long value;
 
     public boolean anyFits(Collection<AchievementEntity> achievements) {
         if (achievements == null || achievements.isEmpty()) {
